+++ flask_unchained/bundles/webpack/extensions/__init__.py
@@ -1,0 +1,9 @@
+from .webpack import Webpack
+
+
+webpack = Webpack()
+
+
+EXTENSIONS = {
+    'webpack': webpack,
+}