+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * never register static assets routes with babel bundle
 * integrate the mail bundle into the `flask_unchained` package
 * integrate the celery bundle into the `flask_unchained` package
+* integrate the webpack bundle into the `flask_unchained` package
 
 ## 0.4.2 (2018/07/21)